+++ src/components/date-picker/pickers/Picker.ts
@@ -7,13 +7,16 @@ import 'moment/locale/nb';
 import 'moment/locale/sv';
 import moment from 'moment/moment';
 import { isAndroidDevice, isIOSDevice } from '../../../util/device';
+import { DateFormatter } from '../dateFormatter';
 
 export abstract class Picker {
+    private dateFormatter: DateFormatter;
+
     protected dateFormat: string;
     protected language: string = 'en';
 
     protected flatpickr: flatpickr.Instance;
-    protected nativePicker;
+    protected nativePicker: boolean;
 
     public constructor(
         dateFormat: string,
@@ -26,9 +29,11 @@ export abstract class Picker {
         if (dateFormat) {
             this.dateFormat = dateFormat;
         }
+        this.dateFormatter = new DateFormatter(language);
 
         this.getWeek = this.getWeek.bind(this);
         this.handleClose = this.handleClose.bind(this);
+        this.handleOnClose = this.handleOnClose.bind(this);
         this.parseDate = this.parseDate.bind(this);
         this.formatDate = this.formatDate.bind(this);
         this.getFlatpickrLang = this.getFlatpickrLang.bind(this);
@@ -36,13 +41,15 @@ export abstract class Picker {
 
     public init(element: HTMLElement, container: HTMLElement, value?: Date) {
         let config: flatpickr.Options.Options = {
-            allowInput: true,
+            clickOpens: this.nativePicker,
             disableMobile: !this.nativePicker,
             formatDate: this.nativePicker ? undefined : this.formatDate,
-            onClose: this.handleClose,
             parseDate: this.nativePicker ? undefined : this.parseDate,
             appendTo: container,
+            onClose: this.handleOnClose,
             defaultDate: value,
+            onValueUpdate: this.handleClose,
+            inline: !this.nativePicker,
             locale:
                 FlatpickrLanguages[this.getFlatpickrLang()] ||
                 FlatpickrLanguages.en,
@@ -58,6 +65,10 @@ export abstract class Picker {
         this.flatpickr = flatpickr(element, config) as flatpickr.Instance; // tslint:disable-line:no-useless-cast
     }
 
+    public redraw() {
+        this.flatpickr.redraw();
+    }
+
     public destroy() {
         if (!this.flatpickr) {
             return;
@@ -71,23 +82,17 @@ export abstract class Picker {
     ): flatpickr.Options.Options;
 
     public formatDate(date: Date) {
-        if (this.nativePicker) {
-            return date ? JSON.stringify(date) : '';
-        }
-        if (date) {
-            return moment(date)
-                .locale(this.getMomentLang())
-                .format(this.dateFormat);
-        }
-        return '';
+        return this.dateFormatter.formatDate(date, this.dateFormat);
     }
 
     protected handleClose(selectedDates): Promise<any> {
-        if (this.nativePicker) {
-            return this.handleCloseForNativePicker(selectedDates);
-        } else {
-            return this.handleCloseForFlatpickr(selectedDates);
-        }
+        return new Promise((resolve) => {
+            setTimeout(() => {
+                const pickerDate = this.getPickerDate(selectedDates);
+                this.change.emit(pickerDate);
+                resolve(pickerDate);
+            }, 0);
+        });
     }
 
     protected getFlatpickrLang() {
@@ -98,50 +103,6 @@ export abstract class Picker {
         return this.language === 'no' ? 'nb' : this.language;
     }
 
-    private handleCloseForNativePicker(selectedDates) {
-        return new Promise((resolve) => {
-            setTimeout(() => {
-                const pickerDate = this.getPickerDate(selectedDates);
-                this.change.emit(pickerDate);
-                resolve(pickerDate);
-            }, 0);
-        });
-    }
-
-    private handleCloseForFlatpickr(selectedDates) {
-        return new Promise((resolve) => {
-            // Since we allow manual editing of the input value, and
-            // flatpickr only picks up these changes when the user presses
-            // enter in the input, we need to check if the input string
-            // and the underlying value match.
-            //
-            // If this timeout is set to 0, we get a race-condition where
-            // the value is sometimes updated from the input-string, and
-            // sometimes not.
-            const timeout = 100;
-            setTimeout(() => {
-                const momentInputDate = moment(
-                    this.flatpickr.input.value,
-                    this.dateFormat,
-                    this.getMomentLang()
-                );
-                let pickerDate = this.getPickerDate(selectedDates);
-                const isSameInput = momentInputDate.isSame(pickerDate);
-                if (!isSameInput) {
-                    if (momentInputDate.isValid()) {
-                        pickerDate = momentInputDate.toDate();
-                        this.flatpickr.setDate(pickerDate);
-                    } else {
-                        pickerDate = null;
-                        this.flatpickr.clear();
-                    }
-                }
-                this.change.emit(pickerDate);
-                resolve(pickerDate);
-            }, timeout);
-        });
-    }
-
     private getPickerDate(selectedDates) {
         return selectedDates[0] ? new Date(selectedDates[0].toJSON()) : null;
     }
@@ -150,11 +111,11 @@ export abstract class Picker {
         return moment(date).isoWeek();
     }
 
-    private parseDate(dateString) {
-        return moment(
-            dateString,
-            this.dateFormat,
-            this.getMomentLang()
-        ).toDate();
+    private parseDate(date: string) {
+        return moment(date, this.dateFormat, this.getMomentLang()).toDate();
+    }
+
+    private handleOnClose() {
+        this.flatpickr.element.focus();
     }
 }